* pre_tag_map: results of evaluation of pre_tag_map, in case of a positive   match, overrides any tags passed by nfprobe/sfprobe plugins via NetFlow/   sFlow export. * fix, plugin_hooks.c: plugin_buffer_size variables are bumped to u_int64_t

+++ src/plugin_hooks.h
@@ -65,9 +65,9 @@ struct aggregate_filter {
 struct channels_list_entry {
   u_int64_t aggregation;
   u_int64_t aggregation_2;
-  u_int32_t buf;	/* buffer base */
-  u_int32_t bufptr;	/* buffer current */
-  u_int32_t bufend;	/* buffer end; max 4Gb */
+  u_int64_t buf;	/* buffer base */
+  u_int64_t bufptr;	/* buffer current */
+  u_int64_t bufend;	/* buffer end; max 4Gb */
   struct ring rg;	
   struct ch_buf_hdr hdr;
   struct ch_status *status;
